requirements.txt
openai>=1.0.0
notion-client>=2.0.0
pymongo>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
